Remove the old 0715, 0122 and 0604 data sets for Chrome.
Adjust scripts accordingly and eliminate old unittest file as well.
As of this revision, the default for compilation in all the scripts
EXCEPT for compile_full.sh is the 20141015 dataset. The compile_full.sh
script still builds the most recent pre-20141015 data set, and should
also be updated soon.

The suffix (0122, 0715, 0604, or 20141015) has simply been deleted from
all affected filenames, e.g.:
  cld2_generated_deltaoctachrome0122.cc -> cld2_generated_deltaoctachrome.cc

Moving forward the intention is not keep the old data files around. Repository
history can be used to check out old versions of data files as necessary.

The clean.sh script has also been updated with the names of the new executables
produced in the compilation and testing process along with the svn:ignore
propset.

In the next commit, the compile20141015 will become compile.sh.



git-svn-id: https://cld2.googlecode.com/svn/trunk@190 b252ecd4-b096-bf77-eb8e-91563289f87e